Add solution #2
Went back through and used the official solution. It is much easier
and more elegant. Used justify-content: space-between to achieve the
desired outcome. Space-between will send the end flex items to the far
ends and keep the middle flex item in the center. Then adjusted the
margin and padding from there.